requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
fastapi==0.114.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
loguru==0.7.2
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
python-telegram-bot==21.5
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
